Refactor: move notoriety component to shared, add prediction, network fields, move event to common
