+++ Content.Goobstation.Client/Wanted/NotorietySystem.cs
@@ -1,0 +1,5 @@
+using Content.Goobstation.Shared.Wanted;
+
+namespace Content.Goobstation.Client.Wanted;
+
+public sealed class NotorietySystem : SharedNotorietySystem;
